refactor(tui): textual-free rust seam
M7 Task 46-pre — the decoupling forced by kickoff recon: six couplings
sat between RustClientHost's import chain and the Textual estate, three
at runtime (campaign_menu -> LobbyScreen widgets, wikilinks -> WIKILINK_RE,
theme -> the KSBC Theme object), two at mypy-time (tutorial_overlay
protocols, directives statblock aliases), plus CampaignHandle & friends
DEFINED INSIDE app.py while every keep-bound test imports them at runtime.

Five extractions, all verbatim moves with re-imports for the dying side:
- babylon/tui/contract.py (NEW): TickOutcome, CampaignHandle,
  CampaignLoader, PacedDriverHandle, DriverFactory,
  TutorialProgressFactory (from app.py) + TutorialStepView,
  TutorialProgress (from tutorial_overlay.py) — the durable client seam.
- babylon/tui/lobby_screen.py (NEW): LobbyScreen widget out of
  campaign_menu.py, which is now textual-free (host.py builds
  CampaignMenu at runtime).
- babylon/tui/wikilink_grammar.py (NEW): WIKILINK_RE + WikilinkResolver +
  known_target_resolver out of wikilinks.py (shell/backlinks.py needs the
  regex textual-free).
- babylon/tui/ksbc_theme.py (NEW): the registered Textual Theme out of
  theme.py, which keeps the palette constants (chronicle/watchlist/peek
  import them at runtime).
- babylon/tui/statblocks.py (NEW): StatblockRow/StatblockProvider out of
  directives.py (projection/organization+institution type-check against
  them).

Re-points: host.py + cli/play.py TYPE_CHECKING blocks, shell/backlinks,
both projection modules, the pilot harness, test_rust_host_m2,
test_tutorial/test_campaign_menu LobbyScreen imports, 7 snapshot apps'
KSBC imports.

ACCEPTANCE (the ceremony's precondition): importing babylon.tui.host +
babylon.cli.play + babylon.tui.contract loads ZERO textual modules —
proven live. Zero behavior change; import-linter 9/9 kept; affected
estates 992+22 green.
